Inline SetupOnlyforOverridableMemebers
