--- Source/Moq.Analyzers.Test/Data/SetupOnlyForOverridableMembers.TestOkForInterfaceProperty.cs
@@ -1,17 +1,0 @@
-﻿using Moq;
-
-namespace SetupOnlyForOverridableMembers.TestOkForInterfaceProperty;
-
-public interface ISampleInterface
-{
-    int TestProperty { get; set; }
-}
-
-internal class MyUnitTests
-{
-    private void TestOkForInterfaceProperty()
-    {
-        var mock = new Mock<ISampleInterface>();
-        mock.Setup(x => x.TestProperty);
-    }
-}
