--- Source/Moq.Analyzers.Test/Data/SetupOnlyForOverridableMembers.TestOkForOverrideAbstractMethod.cs
@@ -1,29 +1,0 @@
-﻿using Moq;
-
-namespace SetupOnlyForOverridableMembers.TestOkForOverrideAbstractMethod;
-
-public abstract class BaseSampleClass
-{
-    public int Calculate() => 0;
-
-    public abstract int Calculate(int a, int b);
-
-    public abstract int Calculate(int a, int b, int c);
-}
-
-public class SampleClass : BaseSampleClass
-{
-
-    public override int Calculate(int a, int b) => 0;
-
-    public sealed override int Calculate(int a, int b, int c) => 0;
-}
-
-internal class MyUnitTests
-{
-    private void TestOkForOverrideAbstractMethod()
-    {
-        var mock = new Mock<SampleClass>();
-        mock.Setup(x => x.Calculate(It.IsAny<int>(), It.IsAny<int>()));
-    }
-}
