--- Source/Moq.Analyzers.Test/Data/SetupOnlyForOverridableMembers.TestOkForVirtualMethod.cs
@@ -1,17 +1,0 @@
-﻿using Moq;
-
-namespace SetupOnlyForOverridableMembers.TestOkForVirtualMethod;
-
-public class SampleClass
-{
-    public virtual int DoSth() => 0;
-}
-
-internal class MyUnitTests
-{
-    private void TestOkForVirtualMethod()
-    {
-        var mock = new Mock<SampleClass>();
-        mock.Setup(x => x.DoSth());
-    }
-}
